move auth fetches to library/api
